Move domTestRunner to test-utils
