+++ test/dom/bindings.mocha.js
@@ -1,8 +1,9 @@
 var expect = require('chai').expect;
-var domTestRunner = require('../domTestRunner');
+var domTestRunner = require('../../test-utils/domTestRunner');
 
 describe('bindings', function() {
   var runner = domTestRunner.createRunner();
+  debugger;
 
   describe('bracket dependencies', function() {
     it('bracket inner dependency change', function() {
